# openapi.yaml
openapi: 3.1.0
info:
  title: Silocat api_switch API
  version: 1.0.0
  description: |
    Internal HTTP API of the silocat `api_switch` service (Rust/axum, listens on :31337,
    exposed as api.silo.cat through Cloudflare and nginx).

    ### Trust model
    This API is service-to-service. Every request must carry the `X-Authority-Sign`
    header whose value equals the server's `AUTHORITY_SIGN` secret; only silocat's own
    backends (web_server SSR, the clickswave admin panel) hold it. A missing or wrong header yields
    `401 {"status":401,"message":"Unauthorized Origin","errors":[],"data":{}}` on every path.

    On top of that, a subset of `/user/*` routes additionally require the end user's
    `X-Api-Key` header (the `users.api_key` value). Banned accounts get 403 from that
    middleware. Most `/file/*`, `/folder/*` and `/billing/*` routes instead take
    `user_id` or `api_key` fields in the request payload and enforce ownership ad hoc;
    the calling backend is trusted to pass authentic values.

    ### Response envelope
    Unless noted otherwise, every response (success and error) is JSON of the shape
    `{status, message, errors, data}` where `status` mirrors the HTTP status code
    (mapped codes: 200, 201, 400, 401, 403, 404, 409, 413, 429, 500; anything else is
    sent as HTTP 500). `data` is `{}` or `null` on most errors. Exceptions:
    `GET /admin/files/{id}/download` streams raw bytes on success, and malformed JSON
...

    ### Uploads and downloads
...
servers:
...
security:
  - AuthoritySign: []
...
paths:
...
  /file/restore-files:
    post:
      tags: [file]
      operationId: fileRestoreFile
      summary: Restore a soft-deleted file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileKeyOwnershipPayload'
      responses:
        '200':
          description: File restored successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '403':
          $ref: '#/components/responses/OwnershipMismatch'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /file/list-files:
    post:
      tags: [file]
      operationId: fileListFiles
      summary: List a user's files (folder view, starred, shared, or trash)
      description: |
        Filter precedence: starred=true, then shared=true (share_type != 'off'), then
        deleted=true (flat trash list, folder_id ignored), else the normal folder view
        (folder_id null means root). Ordered newest first. user_id is trusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: {type: string}
                folder_id: {type: ['string', 'null']}
                starred: {type: ['boolean', 'null']}
                shared: {type: ['boolean', 'null']}
                deleted: {type: ['boolean', 'null']}
      responses:
        '200':
          description: Files retrieved successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          files:
                            type: array
                            items:
                              $ref: '#/components/schemas/File'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /file/star/folder:
    post:
      tags: [file]
      operationId: fileStarFolder
      summary: Star or unstar a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder_id, user_id, starred]
              properties:
                folder_id: {type: string}
                user_id: {type: string}
                starred: {type: boolean}
      responses:
        '200':
          description: Folder star status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /folder/permanent-delete:
    post:
      tags: [folder]
      operationId: folderPermanentDelete
      summary: Permanently delete a folder (children cascade in the database)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder_id, api_key]
              properties:
                folder_id: {type: string}
                api_key:
                  type: string
                  description: Must match the folder's owner_api_key or belong to its owning user.
      responses:
        '200':
          description: Folder permanently deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '403':
          $ref: '#/components/responses/OwnershipMismatch'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
  /folder/restore:
    post:
      tags: [folder]
      operationId: folderRestore
      summary: Restore a soft-deleted folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder_id, api_key]
              properties:
                folder_id: {type: string}
                api_key: {type: string}
      responses:
        '200':
          description: Folder restored successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '403':
          $ref: '#/components/responses/OwnershipMismatch'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/ip-bans:
    get:
      tags: [admin]
      operationId: adminListIpBans
      summary: List IP bans (newest first, max 500)
      responses:
        '200':
          description: IP bans fetched.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          bans:
                            type: array
                            items:
                              $ref: '#/components/schemas/IpBan'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [admin]
      operationId: adminCreateIpBan
      summary: Ban an IP address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ip]
              properties:
                ip: {type: string}
                reason: {type: ['string', 'null']}
                days:
                  type: ['integer', 'null']
                  format: int64
                  description: Days until expiry; null, 0 or negative means permanent.
      responses:
        '200':
          description: IP banned. data.id is the new ban's UUID.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id: {type: string}
        '400':
          description: IP is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/ip-bans/{id}:
    delete:
      tags: [admin]
      operationId: adminDeleteIpBan
      summary: Remove an IP ban (200 even if it did not exist)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: IP ban removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/users/{id}/ban:
    post:
      tags: [admin]
      operationId: adminBanUser
      summary: Ban a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: ['string', 'null']}
                days:
                  type: ['integer', 'null']
                  format: int64
                  description: Days until the ban lifts; null, 0 or negative means permanent.
      responses:
        '200':
          description: User banned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/UnauthorizedOrigin'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    AuthoritySign:
      type: apiKey
      in: header
      name: X-Authority-Sign
      description: |
        Service-to-service secret shared with silocat's own backends (env
        AUTHORITY_SIGN). Required on every request; failure is 401 "Unauthorized
        Origin".
    UserApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        The end user's api_key (users.api_key). Required on protected /user routes in
        addition to X-Authority-Sign. Invalid/missing keys get 401 "Unauthorized";
        banned accounts get 403 "Account banned".
  responses:
    UnauthorizedOrigin:
      description: Missing or invalid X-Authority-Sign header ("Unauthorized Origin").
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
...
    OwnershipMismatch:
      description: The supplied api_key does not match the resource owner.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
    NotFound:
      description: Resource not found (or not owned by the given user).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
...
    ServerError:
      description: Internal or database error. errors[] usually carries the underlying error string.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
  schemas:
    Envelope:
      type: object
      description: |
        Standard response envelope. status mirrors the HTTP status code; errors is a
        list of human-readable strings; data varies per endpoint and is {} or null on
        most errors.
      required: [status, message, errors, data]
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        data:
          description: Endpoint-specific payload.
...
    File:
      type: object
      description: |
        Full files row. Note this includes owner_api_key and share_password_hash;
        callers should strip them before exposing data to end users.
      properties:
        id: {type: string}
        user_id: {type: ['string', 'null']}
        name: {type: string}
        mime: {type: string}
        size: {type: integer, format: int64}
        encrypted: {type: boolean}
        downloads: {type: integer, format: int64}
        created_on: {type: string, format: date-time}
        total_chunks: {type: integer, format: int64}
        uploaded_chunks: {type: integer, format: int64}
        deleted: {type: boolean}
        sha256_checksum: {type: string}
        blake3_checksum: {type: string}
        public_access: {type: boolean}
        folder_id: {type: ['string', 'null']}
        owner_api_key: {type: ['string', 'null']}
        starred: {type: boolean}
        share_token: {type: ['string', 'null']}
        share_type:
          type: ['string', 'null']
          description: off, public or once (by convention).
        link_downloads: {type: ['integer', 'null'], format: int64}
        link_max_downloads: {type: ['integer', 'null'], format: int64}
        share_expires_at: {type: ['string', 'null'], format: date-time}
        share_password_hash: {type: ['string', 'null']}
        deleted_on: {type: ['string', 'null'], format: date-time}
...
    IpBan:
      type: object
      properties:
        id: {type: string}
        ip: {type: string}
        reason: {type: ['string', 'null']}
        banned_until:
          type: ['string', 'null']
          format: date-time
          description: Null means permanent.
        created_at: {type: string, format: date-time}
        permanent: {type: boolean}
        active: {type: boolean}
...
    FileKeyOwnershipPayload:
      type: object
      required: [file_id, api_key]
      properties:
        file_id:
          type: string
        api_key:
          type: string
          description: Must match the file's owner_api_key or belong to the user who owns the file.